puts

+++ 0x05-pointers_arrays_strings/3-puts.c
@@ -9,9 +9,10 @@
 void _puts(char *str)
 {
 	int i = 0;
+
 	while (*str++)
 	{
-		_putchar(*(str + i));
+		_putchar((int) *(str + i));
 		i++;
 	}
 }
